log to ln
The right equation accounts for ln() instead of log(). Using log can increase erros when comparing with other FLIR softwares for temperature calibration.

You can see that also on PR1/(PR2*(exp(PB/(IRWTemp+273.15))-PF))-PO and other similar equations. Here it is used the exp() so natural base. That means that on the temperature calculation the natural logarithm should be used.